+++ src/main/java/domain/board/BoardGenerator.java
@@ -5,14 +5,12 @@ import domain.piece.ChoByeong;
 import domain.piece.HanByeong;
 import domain.piece.Ma;
 import domain.piece.Piece;
-import domain.PieceType;
 import domain.piece.Po;
 import domain.piece.Sa;
 import domain.piece.Sang;
-import domain.Team;
 import domain.piece.Wang;
-import domain.point.PathFinder;
-import domain.point.PathFinderFactory;
+import domain.piece.character.PieceType;
+import domain.piece.character.Team;
 import domain.point.Point;
 import java.util.ArrayDeque;
 import java.util.Deque;
