+++ src/main/java/domain/piece/ObstacleSensitivePiece.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Team;
+import domain.piece.character.Team;
 import domain.board.PieceVisibleBoard;
 import domain.point.Movement;
 import domain.point.Point;
