+++ src/main/java/domain/piece/Sang.java
@@ -21,8 +21,8 @@ import static domain.point.Path.UP_UP_RIGHT_PATH;
 import static domain.point.Path.UP_UP_UP_LEFT_LEFT_PATH;
 import static domain.point.Path.UP_UP_UP_RIGHT_RIGHT_PATH;
 
-import domain.PieceType;
-import domain.Team;
+import domain.piece.character.PieceType;
+import domain.piece.character.Team;
 import domain.point.Movement;
 import java.util.List;
