+++ src/main/java/domain/piece/character/PieceType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece.character;
 
 public enum PieceType {
