+++ src/main/java/domain/piece/character/Team.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece.character;
 
 public enum Team {
