+++ src/test/java/domain/board/PathFinderTest.java
@@ -2,7 +2,6 @@ package domain.board;
 
 import domain.point.Direction;
 import domain.point.Path;
-import domain.point.PathFinder;
 import domain.point.Point;
 import fixture.BoardFixture;
 import java.util.stream.Stream;
