+++ src/test/java/domain/piece/SaTest.java
@@ -1,7 +1,7 @@
 package domain.piece;
 
-import domain.PieceType;
-import domain.Team;
+import domain.piece.character.PieceType;
+import domain.piece.character.Team;
 import domain.board.Board;
 import domain.point.Point;
 import fixture.BoardFixture;
